+++ src/main/java/com/sparcs/teamf/api/auth/exception/RefreshTokenValidationException.java
@@ -1,4 +1,4 @@
-package com.sparcs.teamf.api.auth.error;
+package com.sparcs.teamf.api.auth.exception;
 
 import com.sparcs.teamf.api.error.exception.HttpException;
 import org.springframework.http.HttpStatus;
